move duplicated logic into utils file
