+++ frontend/src/metabase/visualizations/visualizations/SmartScalar.jsx
@@ -18,6 +18,7 @@ import ScalarValue, {
 import { isDate } from "metabase-lib/types/utils/isa";
 import { formatBucketing } from "metabase-lib/queries/utils/query-time";
 import { ScalarContainer } from "./Scalar.styled";
+import { compactifyValue } from "./scalarUtils";
 
 import {
   PreviousValueContainer,
@@ -25,11 +26,6 @@ import {
   PreviousValueVariation,
   Variation,
 } from "./SmartScalar.styled";
-
-// used below to determine whether we show compact formatting
-const COMPACT_MAX_WIDTH = 250;
-const COMPACT_WIDTH_PER_DIGIT = 25;
-const COMPACT_MIN_LENGTH = 6;
 
 export default class Smart extends React.Component {
   static uiName = t`Trend`;
@@ -125,21 +121,11 @@ export default class Smart extends React.Component {
     const lastValue = insight["last-value"];
     const formatOptions = settings.column(column);
 
-    const fullScalarValue = formatValue(lastValue, formatOptions);
-    const compactScalarValue = formatValue(lastValue, {
-      ...formatOptions,
-      compact: true,
+    const { displayValue, fullScalarValue } = compactifyValue({
+      formatOptions,
+      value: lastValue,
+      width,
     });
-
-    // use the compact version of formatting if the component is narrower than
-    // the cutoff and the formatted value is longer than the cutoff
-    // also if the width is less than a certain multiplier of the number of digits
-    const displayCompact =
-      fullScalarValue !== null &&
-      fullScalarValue.length > COMPACT_MIN_LENGTH &&
-      (width < COMPACT_MAX_WIDTH ||
-        width < COMPACT_WIDTH_PER_DIGIT * fullScalarValue.length);
-    const displayValue = displayCompact ? compactScalarValue : fullScalarValue;
 
     const granularity = formatBucketing(insight["unit"]).toLowerCase();
